Massive type cleanup
Using an enum (currently BaseValue) to indicate places in the layout that are supposed to have expressions instead of the fancy ExpressionOr<...> type that allowed for concrete values AND expressions. Now, the ILayoutComponent type itself does not support concrete values, expressions, or indicates that expressions are resolved. To do that, you have to wrap the type in ExprResolved<...> or ExprUnresolved<...> (the latter is deprecated, and we're hereby shifting the complexity of using non-resolved component definitions over to the older deprecated code). This makes working with the newer hierarchy system cleaner and easier than working with legacy code, hopefully increasing adoption and making for better developer experience.